+++ test/object-store/sync/sync_test_utils.cpp
@@ -476,11 +476,189 @@ private:
     TestAppSession& m_test_app_session;
 };
 
+struct BaasFLXClientReset : public TestClientReset {
+    BaasFLXClientReset(const Realm::Config& local_config, const Realm::Config& remote_config,
+                       const TestAppSession& test_app_session)
+        : TestClientReset(local_config, remote_config)
+        , m_test_app_session(test_app_session)
+    {
+        REALM_ASSERT(m_local_config.sync_config->flx_sync_requested);
+        REALM_ASSERT(m_remote_config.sync_config->flx_sync_requested);
+        REALM_ASSERT(m_local_config.schema->find(c_object_schema_name) != m_local_config.schema->end());
+    }
+
+    void wait_for_initial_sync()
+    {
+        // Reduce the liklihood of getting ProtocolError 229:
+        // "attempted to start a session while initial sync is in progress"
+        // This is not a fatal error, and sync should retry within 5 minutes,
+        // but rather not prolong the tests if it can be avoided.
+        // TODO: it should be possible to poll MongoDB to see if it is done
+        // initial bootstraping by checking via the Admin API if this database
+        // name is in the QBS metadata.
+        // https://github.com/10gen/baas/blob/a63be65c69bbeca78a1c31201e7aeee472d120b3/etc/sync/e2e/net/Utils/TestHarness.cs#L313
+        millisleep(2000);
+    }
+
+    void run() override
+    {
+        m_did_run = true;
+        const AppSession& app_session = m_test_app_session.app_session();
+
+        wait_for_initial_sync();
+        auto realm = Realm::get_shared_realm(m_local_config);
+        auto session = realm->sync_session();
+        const ObjectId pk_of_added_object("123456789000000000000000");
+        {
+            if (m_on_setup) {
+                m_on_setup(realm);
+            }
+            constexpr bool create_object = true;
+            subscribe_to_object_by_id(realm, pk_of_added_object, create_object);
+
+            // While at this point the object has been sync'd successfully, we must also
+            // wait for it to appear in the backing database before terminating sync
+            // otherwise the translator may be terminated before it has a chance to
+            // integrate it into the backing database. If the server were to change
+            // the meaning of "upload complete" to include writing to atlas then this would
+            // not be necessary.
+            app::MongoClient remote_client = m_local_config.sync_config->user->mongo_client("BackingDB");
+            app::MongoDatabase db = remote_client.db(app_session.config.mongo_dbname);
+            app::MongoCollection object_coll = db[std::string(c_object_schema_name)];
+            uint64_t count_external = 0;
+
+            timed_sleeping_wait_for(
+                [&]() -> bool {
+                    if (count_external == 0) {
+                        object_coll.count({{std::string(c_id_col_name), pk_of_added_object}},
+                                          [&](uint64_t count, util::Optional<app::AppError> error) {
+                                              REQUIRE(!error);
+                                              count_external = count;
+                                          });
+                    }
+                    if (count_external == 0) {
+                        millisleep(2000); // don't spam the server too much
+                    }
+                    return count_external > 0;
+                },
+                std::chrono::minutes(5));
+            session->log_out();
+
+            if (m_make_local_changes) {
+                m_make_local_changes(realm);
+            }
+        }
+
+        // cause a client reset by restarting the sync service
+        // this causes the server's sync history to be resynthesized
+        auto baas_sync_service = app_session.admin_api.get_sync_service(app_session.server_app_id);
+        auto baas_sync_config = app_session.admin_api.get_config(app_session.server_app_id, baas_sync_service);
+        REQUIRE(app_session.admin_api.is_sync_enabled(app_session.server_app_id));
+        app_session.admin_api.disable_sync(app_session.server_app_id, baas_sync_service.id, baas_sync_config);
+        REQUIRE(!app_session.admin_api.is_sync_enabled(app_session.server_app_id));
+        app_session.admin_api.enable_sync(app_session.server_app_id, baas_sync_service.id, baas_sync_config);
+        REQUIRE(app_session.admin_api.is_sync_enabled(app_session.server_app_id));
+        if (app_session.config.dev_mode_enabled) { // dev mode is not sticky across a reset
+            app_session.admin_api.set_development_mode_to(app_session.server_app_id, true);
+        }
+
+        {
+            wait_for_initial_sync();
+            auto realm2 = Realm::get_shared_realm(m_remote_config);
+            wait_for_download(*realm2);
+            load_initial_data(realm2);
+
+            timed_sleeping_wait_for(
+                [&]() -> bool {
+                    realm2->begin_transaction();
+                    auto table = get_table(*realm2, c_object_schema_name);
+                    auto objkey = table->find_primary_key({pk_of_added_object});
+                    realm2->cancel_transaction();
+                    return bool(objkey);
+                },
+                std::chrono::seconds(60));
+
+            // expect the last sync'd object to be in place
+            realm2->begin_transaction();
+            auto table = get_table(*realm2, c_object_schema_name);
+            REQUIRE(table->size() >= 1);
+            auto obj = table->get_object_with_primary_key({pk_of_added_object});
+            REQUIRE(obj.is_valid());
+            realm2->commit_transaction();
+
+            if (m_make_remote_changes) {
+                m_make_remote_changes(realm2);
+            }
+            wait_for_upload(*realm2);
+            auto subs = realm2->get_latest_subscription_set();
+            subs.get_state_change_notification(sync::SubscriptionSet::State::Complete).get();
+            realm2->close();
+        }
+
+        // Resuming sync on the first realm should now result in a client reset
+        session->revive_if_needed();
+        if (m_on_post_local) {
+            m_on_post_local(realm);
+        }
+        auto subs = realm->get_latest_subscription_set();
+        subs.get_state_change_notification(sync::SubscriptionSet::State::Complete).get();
+        if (m_on_post_reset) {
+            m_on_post_reset(realm);
+        }
+    }
+
+private:
+    void subscribe_to_object_by_id(SharedRealm realm, ObjectId pk, bool create_object = false)
+    {
+        auto mut_subs = realm->get_latest_subscription_set().make_mutable_copy();
+        Group::TableNameBuffer buffer;
+        auto class_name = Group::class_name_to_table_name(c_object_schema_name, buffer);
+        TableRef table = realm->read_group().get_table(class_name);
+        REALM_ASSERT(table);
+        ColKey id_col = table->get_column_key(c_id_col_name);
+        REALM_ASSERT(id_col);
+        ColKey str_col = table->get_column_key(c_str_col_name);
+        REALM_ASSERT(str_col);
+        Query query_for_added_object = table->where().equal(id_col, pk);
+        mut_subs.insert_or_assign(query_for_added_object);
+        auto subs = std::move(mut_subs).commit();
+        if (create_object) {
+            realm->begin_transaction();
+            table->create_object_with_primary_key(pk, {{str_col, "initial value"}});
+            realm->commit_transaction();
+        }
+        subs.get_state_change_notification(sync::SubscriptionSet::State::Complete).get();
+    }
+
+    void load_initial_data(SharedRealm realm)
+    {
+        auto mut_subs = realm->get_latest_subscription_set().make_mutable_copy();
+        for (const auto& table : realm->schema()) {
+            Query query_for_table(realm->read_group().get_table(table.table_key));
+            mut_subs.insert_or_assign(query_for_table);
+        }
+        auto subs = std::move(mut_subs).commit();
+        subs.get_state_change_notification(sync::SubscriptionSet::State::Complete).get();
+    }
+
+    const TestAppSession& m_test_app_session;
+    constexpr static std::string_view c_object_schema_name = "TopLevel";
+    constexpr static std::string_view c_id_col_name = "_id";
+    constexpr static std::string_view c_str_col_name = "queryable_str_field";
+};
+
 std::unique_ptr<TestClientReset> make_baas_client_reset(const Realm::Config& local_config,
                                                         const Realm::Config& remote_config,
                                                         TestAppSession& test_app_session)
 {
     return std::make_unique<BaasClientReset>(local_config, remote_config, test_app_session);
+}
+
+std::unique_ptr<TestClientReset> make_baas_flx_client_reset(const Realm::Config& local_config,
+                                                            const Realm::Config& remote_config,
+                                                            const TestAppSession& session)
+{
+    return std::make_unique<BaasFLXClientReset>(local_config, remote_config, session);
 }
 
 #endif // REALM_ENABLE_AUTH_TESTS